[구독] v0.22 오늘의 식단
식단 추천 시 권장 칼로리 조건 추가

+++ BbsSample/app/src/main/java/com/metamom/bbssample/subscribe/SubDietMealDto.kt
@@ -5,11 +5,11 @@ import android.os.Parcelable
 
 /* #21# 구독 _오늘의 식단 [다이어트] RecyclerView에 드로잉하기 위하여 파일 사용 */
 // subdfType = 타입
-class SubDietMealDto(val subdfSeq:Int, val subdfTime:Int, val subdfImage:String?, val subdfName:String?, val subdfKcal:Int, val subdfAmount:Int, val subdfType:String?) :Parcelable {
+class SubDietMealDto(val subdfSeq:Int, val subdfTime:Int, val subdfImage:String?, val subdfName:String?, val subdfKcal:Double, val subdfAmount:Int, val subdfType:String?) :Parcelable {
 
     // parcel에 대한 기본생성자
     constructor(parcel: Parcel) :this(
-        parcel.readInt(), parcel.readInt(), parcel.readString(), parcel.readString(), parcel.readInt(), parcel.readInt(), parcel.readString()
+        parcel.readInt(), parcel.readInt(), parcel.readString(), parcel.readString(), parcel.readDouble(), parcel.readInt(), parcel.readString()
     ){ }
 
     override fun describeContents(): Int {
@@ -22,7 +22,7 @@ class SubDietMealDto(val subdfSeq:Int, val subdfTime:Int, val subdfImage:String?
         parcel.writeInt(subdfTime)
         parcel.writeString(subdfImage)
         parcel.writeString(subdfName)
-        parcel.writeInt(subdfKcal)
+        parcel.writeDouble(subdfKcal)
         parcel.writeInt(subdfAmount)
         parcel.writeString(subdfType)
     }
