+++ BbsSample/app/src/main/java/com/metamom/bbssample/subscribe/SubExerMealDto.kt
@@ -4,11 +4,11 @@ import android.os.Parcel
 import android.os.Parcelable
 
 /* #21# 구독 _오늘의 식단 [운동] RecyclerView에 드로잉하기 위하여 파일 사용 */
-class SubExerMealDto(val subefSeq:Int, val subefTime:Int, val subefImage:String?, val subefName:String?, val subefKcal:Int, val subefAmount:Int, val subefType:String?) : Parcelable {
+class SubExerMealDto(val subefSeq:Int, val subefTime:Int, val subefImage:String?, val subefName:String?, val subefKcal:Double, val subefAmount:Int, val subefType:String?) : Parcelable {
 
     // parcel에 대한 기본생성자
     constructor(parcel: Parcel) :this(
-        parcel.readInt(), parcel.readInt(), parcel.readString(), parcel.readString(), parcel.readInt(), parcel.readInt(), parcel.readString()
+        parcel.readInt(), parcel.readInt(), parcel.readString(), parcel.readString(), parcel.readDouble(), parcel.readInt(), parcel.readString()
     ){ }
 
     override fun describeContents(): Int {
@@ -21,7 +21,7 @@ class SubExerMealDto(val subefSeq:Int, val subefTime:Int, val subefImage:String?
         parcel.writeInt(subefTime)
         parcel.writeString(subefImage)
         parcel.writeString(subefName)
-        parcel.writeInt(subefKcal)
+        parcel.writeDouble(subefKcal)
         parcel.writeInt(subefAmount)
         parcel.writeString(subefType)
     }
